requirements.txt
# MCP SDK (from GitHub)
mcp @ git+https://github.com/modelcontextprotocol/python-sdk.git

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI Providers
openai>=1.3.0
anthropic>=0.7.0
groq>=0.4.0

# Platform APIs
PyGithub>=2.1.1
python-gitlab>=4.2.0
atlassian-python-api>=3.41.0  # Bitbucket
azure-devops>=7.1.0b4

# Utilities
httpx>=0.25.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
gitpython>=3.1.40

# Diff & Code Analysis
unidiff>=0.7.5
pygments>=2.17.0

# Logging & Monitoring
structlog>=23.2.0

# Security
cryptography>=41.0.7
pyjwt>=2.8.0

